+++ bika/lims/content/samplepoint.py
@@ -141,26 +141,14 @@ class SamplePoint(BaseContent, HistoryAwareMixin, ClientAwareMixin):
     def Title(self):
         return safe_unicode(self.getField('title').get(self)).encode('utf-8')
 
-    def getSampleTypeTitles(self):
-        """Returns a list of sample type titles
-        """
-        sample_types = self.getSampleTypes()
-        sample_type_titles = map(lambda obj: obj.Title(), sample_types)
-
-        # N.B. This is used only for search purpose, because the catalog does
-        #      not add an entry to the Keywordindex for an empty list.
-        #
-        # => This "empty" category allows to search for values with a certain
-        #    sample type set OR with no sample type set.
-        #    (see bika.lims.browser.analysisrequest.add2.get_sampletype_info)
-        if not sample_type_titles:
-            return [""]
-        return sample_type_titles
-
     def getSampleTypeTitle(self):
         """Returns a comma separated list of sample type titles
         """
-        return ",".join(self.getSampleTypeTitles())
+        sample_types = self.getSampleTypes()
+        sample_type_titles = map(lambda obj: obj.Title(), sample_types)
+        if not sample_type_titles:
+            return ""
+        return ",".join(sample_type_titles)
 
     def SampleTypesVocabulary(self):
         from bika.lims.content.sampletype import SampleTypes