+++ CHANGES.rst
@@ -7,6 +7,7 @@ Changelog
 
 **Added**
 
+- #1481 Filter Templates field when Sample Type is selected in Sample Add form
 - #1466 Support for "readonly" and "hidden" visibility modes in ReferenceWidget